cr feedback, refactored graph_for_module to the graph controller -- where it should belong
